Fixed a major bug in the ParksMcClellan code.
The PM code treats the arrays as if they were 1-based, not 0-based.  The only reason my unit test was working was because I had forgotten to take some code out of the PM code that sets the parameters, and the parameters happened to be the same as in the unit test.  Embarrassing.

Also increased the temporary array sizes by 1 because sometimes there were memory errors and I believe it was caused by the 1-based thing.